move almost everything into build.xml
I've moved the contents of buildC#CPP.xml, buildJars.xml, and
buildSetup.xml to build.xml.  There's a lot of rearranging to do
as well, but this is a pretty good start.  The Java, C++ and C#
code all compiles, as far as I know.

If this goes okay, the three soon-to-be-deprecated XML files will
be removed from source control.